Add function to fetch city data
The getWeather function returns the generic city data, which in turn calls on the getWeatherData function to fetch the weather information object.
Also added a function to get the current user location which can be used to populate the page with data before a search is carried out. This was achieved using the built-in geolocation api.